+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Allow comments on `json` configs.
 - Improve MCP tool call feedback.
+- Add support for env vars in mcp configs.
 
 ## 0.0.4